requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
pydantic-settings
PyJWT[crypto]
python-multipart
python-dotenv
httpx
yfinance